Moveout UI layer methods to new interface.
