+++ src/OrchardCore.Modules/OrchardCore.ContentTypes/Controllers/AdminController.cs
@@ -21,6 +21,7 @@ namespace OrchardCore.ContentTypes.Controllers;
 public sealed class AdminController : Controller
 {
     private readonly IContentDefinitionService _contentDefinitionService;
+    private readonly IContentDefinitionViewModelService _contentDefinitionAppService;
     private readonly IContentDefinitionManager _contentDefinitionManager;
     private readonly IAuthorizationService _authorizationService;
     private readonly IDocumentStore _documentStore;
@@ -34,6 +35,7 @@ public sealed class AdminController : Controller
     public AdminController(
         IContentDefinitionDisplayManager contentDefinitionDisplayManager,
         IContentDefinitionService contentDefinitionService,
+        IContentDefinitionViewModelService contentDefinitionAppService,
         IContentDefinitionManager contentDefinitionManager,
         IAuthorizationService authorizationService,
         IDocumentStore documentStore,
@@ -52,6 +54,7 @@ public sealed class AdminController : Controller
 
         H = htmlLocalizer;
         S = stringLocalizer;
+        _contentDefinitionAppService = contentDefinitionAppService;
     }
 
     public Task<ActionResult> Index()
@@ -71,7 +74,7 @@ public sealed class AdminController : Controller
 
         return View("List", new ListContentTypesViewModel
         {
-            Types = await _contentDefinitionService.GetTypesAsync(),
+            Types = await _contentDefinitionAppService.GetTypesAsync(),
         });
     }
 
@@ -101,7 +104,7 @@ public sealed class AdminController : Controller
         {
             ModelState.AddModelError("DisplayName", S["The Display Name can't be empty."]);
         }
-        var types = await _contentDefinitionService.LoadTypesAsync();
+        var types = await _contentDefinitionAppService.LoadTypesAsync();
 
         if (types.Any(t => string.Equals(t.DisplayName.Trim(), viewModel.DisplayName.Trim(), StringComparison.OrdinalIgnoreCase)))
         {
@@ -156,7 +159,7 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var typeViewModel = await _contentDefinitionService.GetTypeAsync(id);
+        var typeViewModel = await _contentDefinitionAppService.GetTypeAsync(id);
 
         if (typeViewModel == null)
         {
@@ -222,7 +225,7 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var typeViewModel = await _contentDefinitionService.LoadTypeAsync(id);
+        var typeViewModel = await _contentDefinitionAppService.LoadTypeAsync(id);
 
         if (typeViewModel == null)
         {
@@ -244,7 +247,7 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var typeViewModel = await _contentDefinitionService.GetTypeAsync(id);
+        var typeViewModel = await _contentDefinitionAppService.GetTypeAsync(id);
 
         if (typeViewModel == null)
         {
@@ -256,7 +259,7 @@ public sealed class AdminController : Controller
         var viewModel = new AddPartsViewModel
         {
             Type = typeViewModel,
-            PartSelections = (await _contentDefinitionService.GetPartsAsync(metadataPartsOnly: false))
+            PartSelections = (await _contentDefinitionAppService.GetPartsAsync(metadataPartsOnly: false))
                 .Where(cpd => !typePartNames.Contains(cpd.Name, StringComparer.OrdinalIgnoreCase) && cpd.PartDefinition != null && cpd.PartDefinition.GetSettings<ContentPartSettings>().Attachable)
                 .Select(cpd => new PartSelectionViewModel { PartName = cpd.Name, PartDisplayName = cpd.DisplayName, PartDescription = cpd.Description })
                 .ToList(),
@@ -273,14 +276,14 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var typeViewModel = await _contentDefinitionService.GetTypeAsync(id);
+        var typeViewModel = await _contentDefinitionAppService.GetTypeAsync(id);
 
         if (typeViewModel == null)
         {
             return NotFound();
         }
 
-        var reusableParts = (await _contentDefinitionService.GetPartsAsync(metadataPartsOnly: false))
+        var reusableParts = (await _contentDefinitionAppService.GetPartsAsync(metadataPartsOnly: false))
                 .Where(cpd => cpd.PartDefinition != null &&
                     cpd.PartDefinition.GetSettings<ContentPartSettings>().Attachable &&
                     cpd.PartDefinition.GetSettings<ContentPartSettings>().Reusable);
@@ -305,7 +308,7 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var typeViewModel = await _contentDefinitionService.LoadTypeAsync(id);
+        var typeViewModel = await _contentDefinitionAppService.LoadTypeAsync(id);
 
         if (typeViewModel == null)
         {
@@ -342,7 +345,7 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var typeViewModel = await _contentDefinitionService.LoadTypeAsync(id);
+        var typeViewModel = await _contentDefinitionAppService.LoadTypeAsync(id);
 
         if (typeViewModel == null)
         {
@@ -417,7 +420,7 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var typeViewModel = await _contentDefinitionService.LoadTypeAsync(id);
+        var typeViewModel = await _contentDefinitionAppService.LoadTypeAsync(id);
 
         if (typeViewModel == null)
         {
@@ -453,7 +456,7 @@ public sealed class AdminController : Controller
         return View(new ListContentPartsViewModel
         {
             // Only user-defined parts (not code as they are not configurable).
-            Parts = await _contentDefinitionService.GetPartsAsync(metadataPartsOnly: true),
+            Parts = await _contentDefinitionAppService.GetPartsAsync(metadataPartsOnly: true),
         });
     }
 
@@ -483,7 +486,7 @@ public sealed class AdminController : Controller
             ModelState.AddModelError("Name", S["The Technical Name can't be empty."]);
         }
 
-        if ((await _contentDefinitionService.LoadPartsAsync(false)).Any(p => string.Equals(p.Name.Trim(), viewModel.Name.Trim(), StringComparison.OrdinalIgnoreCase)))
+        if ((await _contentDefinitionAppService.LoadPartsAsync(false)).Any(p => string.Equals(p.Name.Trim(), viewModel.Name.Trim(), StringComparison.OrdinalIgnoreCase)))
         {
             ModelState.AddModelError("Name", S["A part with the same Technical Name already exists."]);
         }
@@ -508,7 +511,7 @@ public sealed class AdminController : Controller
             return View(viewModel);
         }
 
-        var partViewModel = await _contentDefinitionService.AddPartAsync(viewModel);
+        var partViewModel = await _contentDefinitionAppService.AddPartAsync(viewModel);
 
         if (partViewModel == null)
         {
@@ -586,7 +589,7 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var partViewModel = await _contentDefinitionService.LoadPartAsync(id);
+        var partViewModel = await _contentDefinitionAppService.LoadPartAsync(id);
 
         if (partViewModel == null)
         {
@@ -617,7 +620,7 @@ public sealed class AdminController : Controller
             return RedirectToAction(nameof(List));
         }
 
-        var partViewModel = await _contentDefinitionService.LoadPartAsync(id);
+        var partViewModel = await _contentDefinitionAppService.LoadPartAsync(id);
 
         if (partViewModel == null)
         {
@@ -642,7 +645,7 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var partViewModel = await _contentDefinitionService.LoadPartAsync(id);
+        var partViewModel = await _contentDefinitionAppService.LoadPartAsync(id);
 
         if (partViewModel == null)
         {
@@ -724,7 +727,7 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var partViewModel = await _contentDefinitionService.GetPartAsync(id);
+        var partViewModel = await _contentDefinitionAppService.GetPartAsync(id);
 
         if (partViewModel == null)
         {
@@ -767,7 +770,7 @@ public sealed class AdminController : Controller
             return NotFound();
         }
 
-        var partViewModel = await _contentDefinitionService.LoadPartAsync(id);
+        var partViewModel = await _contentDefinitionAppService.LoadPartAsync(id);
 
         if (partViewModel == null)
         {
@@ -793,7 +796,7 @@ public sealed class AdminController : Controller
                 ModelState.AddModelError("DisplayName", S["The Display Name can't be empty."]);
             }
 
-            if ((await _contentDefinitionService.LoadPartAsync(partViewModel.Name)).PartDefinition.Fields.Any(t => t.Name != viewModel.Name && string.Equals(t.DisplayName().Trim(), viewModel.DisplayName.Trim(), StringComparison.OrdinalIgnoreCase)))
+            if ((await _contentDefinitionAppService.LoadPartAsync(partViewModel.Name)).PartDefinition.Fields.Any(t => t.Name != viewModel.Name && string.Equals(t.DisplayName().Trim(), viewModel.DisplayName.Trim(), StringComparison.OrdinalIgnoreCase)))
             {
                 ModelState.AddModelError("DisplayName", S["A field with the same Display Name already exists."]);
             }
@@ -811,7 +814,7 @@ public sealed class AdminController : Controller
             await _notifier.InformationAsync(H["Display name changed to {0}.", viewModel.DisplayName]);
         }
 
-        await _contentDefinitionService.AlterFieldAsync(partViewModel, viewModel);
+        await _contentDefinitionAppService.AlterFieldAsync(partViewModel, viewModel);
 
         // Refresh the local field variable in case it has been altered
         field = (await _contentDefinitionManager.LoadPartDefinitionAsync(id)).Fields.FirstOrDefault(x => string.Equals(x.Name, viewModel.Name, StringComparison.OrdinalIgnoreCase));
@@ -837,7 +840,7 @@ public sealed class AdminController : Controller
         else
         {
             // Redirect to the type editor if a type exists with this name
-            var typeViewModel = await _contentDefinitionService.LoadTypeAsync(id);
+            var typeViewModel = await _contentDefinitionAppService.LoadTypeAsync(id);
             if (typeViewModel != null)
             {
                 return RedirectToAction(nameof(Edit), new { id });
@@ -855,7 +858,7 @@ public sealed class AdminController : Controller
             return Forbid();
         }
 
-        var partViewModel = await _contentDefinitionService.LoadPartAsync(id);
+        var partViewModel = await _contentDefinitionAppService.LoadPartAsync(id);
 
         if (partViewModel == null)
         {
@@ -873,7 +876,7 @@ public sealed class AdminController : Controller
 
         await _notifier.SuccessAsync(H["The \"{0}\" field has been removed.", field.DisplayName()]);
 
-        if (await _contentDefinitionService.LoadTypeAsync(id) != null)
+        if (await _contentDefinitionAppService.LoadTypeAsync(id) != null)
         {
             return RedirectToAction(nameof(Edit), new { id });
         }
@@ -977,7 +980,7 @@ public sealed class AdminController : Controller
             }
         }
 
-        await _contentDefinitionService.AlterTypePartAsync(viewModel);
+        await _contentDefinitionAppService.AlterTypePartAsync(viewModel);
 
         // Refresh the local part variable in case it has been altered
         part = (await _contentDefinitionManager.LoadTypeDefinitionAsync(id)).Parts.FirstOrDefault(x => string.Equals(x.Name, viewModel.Name, StringComparison.OrdinalIgnoreCase));
