+++ src/OrchardCore.Modules/OrchardCore.ContentTypes/Services/IContentDefinitionService.cs
@@ -1,18 +1,9 @@
 using OrchardCore.ContentManagement.Metadata.Models;
-using OrchardCore.ContentTypes.ViewModels;
 
 namespace OrchardCore.ContentTypes.Services;
 
 public interface IContentDefinitionService
 {
-    Task<IEnumerable<EditTypeViewModel>> LoadTypesAsync();
-
-    Task<IEnumerable<EditTypeViewModel>> GetTypesAsync();
-
-    Task<EditTypeViewModel> LoadTypeAsync(string name);
-
-    Task<EditTypeViewModel> GetTypeAsync(string name);
-
     Task<ContentTypeDefinition> AddTypeAsync(string name, string displayName);
 
     Task RemoveTypeAsync(string name, bool deleteContent);
@@ -27,16 +18,6 @@ public interface IContentDefinitionService
 
     Task<string> GenerateFieldNameFromDisplayNameAsync(string partName, string displayName);
 
-    Task<IEnumerable<EditPartViewModel>> LoadPartsAsync(bool metadataPartsOnly);
-
-    Task<IEnumerable<EditPartViewModel>> GetPartsAsync(bool metadataPartsOnly);
-
-    Task<EditPartViewModel> LoadPartAsync(string name);
-
-    Task<EditPartViewModel> GetPartAsync(string name);
-
-    Task<EditPartViewModel> AddPartAsync(CreatePartViewModel partViewModel);
-
     Task RemovePartAsync(string name);
 
     Task<IEnumerable<Type>> GetFieldsAsync();
@@ -46,10 +27,6 @@ public interface IContentDefinitionService
     Task AddFieldToPartAsync(string fieldName, string displayName, string fieldTypeName, string partName);
 
     Task RemoveFieldFromPartAsync(string fieldName, string partName);
-
-    Task AlterFieldAsync(EditPartViewModel partViewModel, EditFieldViewModel fieldViewModel);
-
-    Task AlterTypePartAsync(EditTypePartViewModel partViewModel);
 
     Task AlterTypePartsOrderAsync(ContentTypeDefinition typeDefinition, string[] partNames);
