+++ src/OrchardCore.Modules/OrchardCore.ContentTypes/Services/StereotypeService.cs
@@ -8,17 +8,17 @@ namespace OrchardCore.ContentTypes.Services;
 public class StereotypeService : IStereotypeService
 {
     private readonly IEnumerable<IStereotypesProvider> _providers;
-    private readonly IContentDefinitionService _contentDefinitionService;
+    private readonly IContentDefinitionViewModelService _contentDefinitionAppService;
     private readonly ILogger<StereotypeService> _logger;
 
     public StereotypeService(
         IEnumerable<IStereotypesProvider> providers,
-        IContentDefinitionService contentDefinitionService,
-        ILogger<StereotypeService> logger)
+        ILogger<StereotypeService> logger,
+        IContentDefinitionViewModelService contentDefinitionAppService)
     {
         _providers = providers;
-        _contentDefinitionService = contentDefinitionService;
         _logger = logger;
+        _contentDefinitionAppService = contentDefinitionAppService;
     }
 
     public async Task<IEnumerable<StereotypeDescription>> GetStereotypesAsync()
@@ -28,7 +28,7 @@ public class StereotypeService : IStereotypeService
         var stereotypes = providerStereotypes.Select(providerStereotype => providerStereotype.Stereotype)
             .ToHashSet(StringComparer.OrdinalIgnoreCase);
 
-        foreach (var contentType in await _contentDefinitionService.GetTypesAsync())
+        foreach (var contentType in await _contentDefinitionAppService.GetTypesAsync())
         {
             if (!contentType.TypeDefinition.TryGetStereotype(out var stereotype) ||
                 stereotypes.Contains(stereotype))
