+++ src/OrchardCore.Modules/OrchardCore.DataLocalization/Services/ContentFieldLocalizationDataProvider.cs
@@ -1,6 +1,3 @@
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using OrchardCore.ContentTypes.Services;
 using OrchardCore.Localization.Data;
 
@@ -8,18 +5,18 @@ namespace OrchardCore.DataLocalization.Services;
 
 public class ContentFieldDataLocalizationProvider : ILocalizationDataProvider
 {
-    private readonly IContentDefinitionService _contentDefinitionService;
+    private readonly IContentDefinitionViewModelService _contentDefinitionAppService;
 
     private static readonly string _contentFieldsContext = "Content Fields";
 
-    public ContentFieldDataLocalizationProvider(IContentDefinitionService contentDefinitionService)
+    public ContentFieldDataLocalizationProvider(IContentDefinitionViewModelService contentDefinitionAppService)
     {
-        _contentDefinitionService = contentDefinitionService;
+        _contentDefinitionAppService = contentDefinitionAppService;
     }
 
     // TODO: Check if there's a better way to get the fields
     public async Task<IEnumerable<DataLocalizedString>> GetDescriptorsAsync()
-        => (await _contentDefinitionService.GetTypesAsync())
+        => (await _contentDefinitionAppService.GetTypesAsync())
             .SelectMany(t => t.TypeDefinition.Parts)
             .SelectMany(p => p.PartDefinition.Fields.Select(f => new DataLocalizedString(_contentFieldsContext, f.Name, string.Empty)));
 }
